tests, net, kubectl: Drop status interface order test
The existing e2e test has been covering the correct order of the
reported interfaces in the VMI status.
The order influenced the correct correspondence between the status
report and the interface which relates to the primary network.

This change removes the test from the e2e suite.
The existing unit tests [1] and the low functional impact of this
scenario using `kubectl` merits the reduction in e2e coverage.

[1] https://github.com/kubevirt/kubevirt/blob/release-1.2/pkg/network/setup/netstat_test.go#L586

Signed-off-by: Edward Haas <edwardh@redhat.com>