move handleDataVolumes() from sync to startVMI
Currently, virt-controller will create DataVolumeTemplates once the VM is created. Even if the VM is stopped. It also continually ensures that DataVolumeTemplates exist whenever a VM is reconciled. Even if the VM is stopped. This has a couple drawbacks:

If the VM is never started, the underlying PVC (if on immediate bind StorageClass) was allocated unnecessarily
It is hard for the user or third party backup software to replace a DataVolume's underlying PVC. See Request to not create DV/PVC if the VM is in Stopped and DV is deleted #11637

Moving handlleDataVolumes() to startVMI addresses two issues above

Signed-off-by: Michael Henriksen <mhenriks@redhat.com>