Change storage tests from withPVC to withStorage
WithPVC uses the PVC stanza in the datavolume object
and that doesn't have the benefits of using the
storage stanza. Like determining the volume modes
and access modes based on the storage class. This
causes issues with providers that don't support
certain combinations that are defaulted to with the
withPVC stanza.

This commit changes the tests to use WithStorage
instead so the right modes are determined from the
storage profile.

Signed-off-by: Alexander Wels <awels@redhat.com>