net, hotplug: Mark network hotplug feature as Beta
The network hotplug feature, limited to bridge and SR-IOV binding
interfaces has not changed since v1.1. There has been no rejections from
the community so far and no special additions are foreseen at the moment

The network hotplug feature has a Feature Gate which is now marked as
Beta in the next Kubevirt release (v1.3).

This change also updates the documentation in a retrospective to specify
that it was in Alpha from its introduction in v1.1

Signed-off-by: Edward Haas <edwardh@redhat.com>