libvmi: keep WithRunning functional by unsetting runstrategy
Since we cannot simply abandon running (semver promise), we will keep it in many test suites
to ensure it does not break.
To do that, we have to unset the runstrategy, since they're mutually
exclusive.

While not deprecated, We discourage users from using this over runStrategy,
so this will align libvmi with the real world in that sense.

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>