admission: Set a default random VM UUID
Today the firmware UUID of a VMI is - if unset - generated at runtime
based on the VMI name. This can lead to collisions between namespaces and
between clusters.

This change is starting to set random firmware UUIDs at VM creation time
in order to prevent this problem.

GtOps workflows have to define the UUID in git if a stable UUID is required.

Relates to https://github.com/kubevirt/kubevirt/pull/11199

Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>