mem-hotplug: don't setup virtio-mem at VM startup
This wil allow to change how virtio-mem is used in KubeVirt.

Instead of having a virtio-mem device setup from the VM boot
a virtio-mem device will only be attached to the domain
when hotplug is requested.

This makes sure that VM configurations that are incompatible
with hotplug can start fine when vmRolloutStrategy is set
to `LiveUpdate` and only the hotplug operation will fail
if memory hotplug is not supported by the VM.

This approach is more stable and much safer.

Signed-off-by: Antonio Cardace <acardace@redhat.com>