Follow general recommendation in setting go toolchain version
It is recommended to either specify the toolchain in it's own directive,
or omit the toolchain directive and specify it on the go version string.
Today, we have a bit of a mixture, which may or may not result in
unexpected builds when building kubevirt on different platforms (go 1.22.0 "command" and 1.22.2 toolchain).

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>