tests: Avoid muti user-data sources in cloud-init
The recent refactoring of the cloud-init spec builders have revealed
that several tests have been specifying user-data source and the factory
that they used to create the VMI added others.
Eventually the one specified by the tests override the ones in the
factories, but now special treatment is needed to make this explicit.

The change includes two items:
- Do not use the user-data cloud-init if not needed (the network test).
- Change the VMI builder factories that add user-data cloud-init by
  default (for boot purposes), to add it only if no user-data exists
  after processing the caller building blocks.

Signed-off-by: Edward Haas <edwardh@redhat.com>