controller: Fix panic when setupHotplug is called with instance types
This change moves setGuestMemory and setupHotplug code to after any
instance type has been applied to the VMI. This ensures memory has been
populated and avoids a panic within setupMemoryHotplug.

To allow this to happen without breaking all of our unit tests all VMI
setup code is also moved under setupVMIFromVM. This should also likely
be done on main but given time constraints we are going to do this first
on release-1.2.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>