wu-ctrl: Use libvmi
Replace the NewMinimalVMI with explicit fields manipulation with
the libvmi and options functions.

Also, removed the usage of functions which encapsulate several steps,
in favor of a more clean and explicit behavior:
Previously `newVirtualMachine` creates a vmi, a pod and adds them into
the related informers; now these steps are explicit in the test body,
so that the reader can have an immediate overview of the informers content.

Signed-off-by: fossedihelm <ffossemo@redhat.com>