Change the startAllVMIs lambda to be a regular function
The startAllVMIs lambda in operator.go, can be a regular function.
Defining it as a variable and setting its logic in another place is hard
to read and debug, and is confusing.

This commit change it to be a regular function.

In addition, the function is misleading, because it doesn't just run the
VMI, but also creates them. This commit also rename the function to
`createRunningVMIs`.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>