Drop the usesSha function from operator.go
The usesSha performs a one-line regular check if string contains another
string, and returns boolean. The caller still need to asert this
returned value, using the `BeTrue()` gomega matcher. This gives unclear
error message in case of assertion failure.

This commit removes this function, replacing it with the dedicated
gomega's `ContainSubstring()` matcher, in order to get cleare error
message.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>