monitoring: Fix wrong calculation for KubeVirtVMIExcessiveMigrations in an edge case
There is an edge case, usually during an upgrade, in which two or more different virt-controllers are reporting the 'kubevirt_vmi_migration_succeeded' metric
for the same migration (VMIM) on the same VMI. In that case, we need to count that migration only once regardless of it being reported multiple times
for the purpose of firing the alert about excessive amount of migrations in the last 24 hours.
The new PromQL isolates the entries with different VMIMs, then sets the value of '1' to all of them, and then we're summing by the pair of vmi and its namespace.

Signed-off-by: Oren Cohen <ocohen@redhat.com>