mem-hotplug: enable only for VMs with memory >= 1Gi
The 1GiB mark is chosen as a tradeoff, it is enough
memory for the guest kernel to allocate its internal data
structures and to allocate the SWIOTLB.
It also means we can memory map all PCI devices
as they're memory mapped in the first 1Gi (PCI hole).

If there are 32bit devices that need to do DMA,
the kernel will allocate memory for the SWIOTLB, which is a
bounce buffer to address memory after the 4Gi Mark, the
SWIOTLB in Linux at the moment of this commit is 64MB and
this causes an issue for the kernel with VMs with very small
amounts of memory, if not enough memory is available this can
lead to the guest kernel panicking while trying to allocate memory.

This SWIOTLB issue is present when setting `maxMemory` (which is set in
all VMs when `LiveUpdate` is used as the `vmRolloutStrategy`) in the
XML domain of the VM because the kernel will have to allow DMA to happen
also for virtio-mem hotplugged memory, which is always mapped after the
4Gi mark and thus always needs SWIOTLB for devices that use 32bit
addresses.

Signed-off-by: Antonio Cardace <acardace@redhat.com>
(cherry picked from commit 6a3004ca17190b205cd5d0f762d662e8585f32da)