tests, storage: Expect the added empty-disk to be at `/dev/vdb`
Previously, an implicit cloudinit disk was added for Cirros VMs to
assist with its boot. That disk was added before the explicit ones.

Now, the cloudinit is added at the tail of the disk/volume list in
case no explicit ones have been mentioned in the input.
Therefore, the test is updated to use now `/dev/vdb` instead of
`/dev/vdc`.

Signed-off-by: Edward Haas <edwardh@redhat.com>