network, netbinding: Add logic to calculate memory overhead
There are network binding plugins that require additional
memory overhead to be added to the virt-launcher's compute
container.

Introduce the logic to calculate the required memory overhead
for when using network binding plugins.

The logic will add the memory overhead once, per unique usage
of a network binding plugin.

This logic will be integrated in the following commits.

Signed-off-by: Orel Misan <omisan@redhat.com>