tests/instancetype: reduce RunVMAndExpectLaunchWithRunStrategy usage
RunVMAndExpectLaunchWithRunStrategy is complicated, inefficient, and has
a complicated name to reflect its complexity. This commit sets the
RunStrategy as part of the VM definition, istead of attempting to update
it, and using gomega constracts to wait for the VM to be ready.

One usage of RunVMAndExpectLaunchWithRunStrategy is slightly more
complicated and is left for a future PR.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>