Improve container ready check
The existing ContainerDisksReady is only checking
if the socket exists. This might be in very rare case
not enough. Such case is when the container disk container
is created later and the handler performs the check right
after the socket was created but before it is able to accept
a connection.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>