net e2e, passt: Add memory overhead
Register the passt network binding plugin with 500Mi of
memory overhead for the virt-launcher pod's `compute` container.

This is done in order to address passt's memory requirements
in the worst case - when all ports are being forwarded.

The discontinued core passt binding caused an addition of
`800Mi` of memory to the compute container when all ports
were forwarded [1].

According to passt's maintainer, we could now use `500Mi`.

[1] https://github.com/kubevirt/kubevirt/pull/11915/files#diff-dddda3558b6627363a57e2d9b76ee63b69c44aa6a7b88343626f522795d0ce7aL395

Signed-off-by: Orel Misan <omisan@redhat.com>