Add two methods to return the VMI min memory by arch
Since the architecture can't be changed on runtime, there is no point to
re-calculate the minimal memory each time we're building a new VMI
object.

Instead, added an interface that will provide this information, after
computing it on architecture registration.

This interface have two private implementations, default and arm64. On
registration, the right implementation is choosen.

The added two function to expose this information:

```
mem := libvmifact.GetMinimalMemory()
// OR
mem := libvmifact.GetQemuMinimalMemory()
```

The private `architecture` variable and `isARM64` function are no longer
in use, and were removed.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>