handler: finalize the migration only when it is in complete state
The workload update migration relies on the change condition on the VMI
to understand if the update migration needs to be canceled.
The removal of the condition and the migration set to completed are
racy. This change helps to coordinates the removal of the change
conditions only after the migration it has been set as final.

Signed-off-by: Alice Frosi <afrosi@redhat.com>