Avoid cancelling the migration when target is ready
Make sure that the migration has completed by checking if the
target domain is ready. In certain cases, the migration can be
still in running phase but the migration process has already completed
and the object hasn't been updated yet. In order to prevent the race
condition of deleting the migration object, we check if the target is
ready.

Signed-off-by: Alice Frosi <afrosi@redhat.com>