Network Binding Plugin: Decouple memory overhead calculation
Following PR #12235, it is possible to specify memory overhead
for the compute container when registering a network binding plugin.

Decouple the memory overhead calculation when:
1. Calculating the compute container's memory overhead.
2. Calculating if there is enough memory in a memory hotplug scenario.

Introduce the `netBindingPluginMemoryCalculator` interface
to:
1. templateService
2. VirtualMachineController

Expect the network binding plugin memory overhead calculation
to be called in unit tests.

There is a check in `template.go` to see if the newly introduced interface
is not `nil` in order to not break existing unit tests that use
the real `templateService` struct:
- VMI controller
- Migration controller

Signed-off-by: Orel Misan <omisan@redhat.com>