converter: fix calculation of the amount of virtIO block queues
the issue was that the cpuCount variable was holding the
total amount of CPUs instead of holding the amount of current
active CPUs. Therefore the calculation of block queues could be
incorrect in cases where CPU hotplug is configured.

Added unit test to cover the case when CPU hotplug is configured
on the VMI, and the # of total CPU can be different
than the # of active CPUs.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>