api: add volume migration info to the VM
Since the VMI isn't persistent, if there are any crashes or the VM is
shutdown, we loose the information regarding the volume migration. If
the migration was in progress or not successful, then the VM is in an
inconsistent since it has the destination volumes in the spec, but the
copied hasn't completed.
Adding the migrated volumes in the VM status as well, helps to give a
hint to the users or to the overlaying tools of what happen to the
original state of the volumes. The migrated volumes on the VM status are
updated to the next volume migration. The conditions gives the history
when the migration was started and if it was completed successfully.

Signed-off-by: Alice Frosi <afrosi@redhat.com>