api: add volume migration state to the VM
Since the VMI isn't persistent, if there are any crashes or the VM is
shutdown, we loose the information regarding the volume migration. If
the migration was in progress or not successful, then the VM is in an
inconsistent since it has the destination volumes in the spec, but the
copy hasn't completed.
Adding the information about the volume migration in the VM status as well,
helps to give a hint to the users or to the overlaying tools of what happens
to the original state of the volumes.

The volume migration state contains the information about the migrated
volumes and if the migration wasn't successful, and a manual recovery is
necessary.

Signed-off-by: Alice Frosi <afrosi@redhat.com>