virt-api: Reject Matcher update with Infer, Name and Kind set
This change provides a more meaningful error to the user when updating a
Matcher that already has a Name and a Kind set with InferFromVolume.

This is a common mistake users can make when accidentally reapplying a
manifest to an existing VM. Ultimately we shouldn't be mutating the Spec
of the VM and thus causing this but for now this commit just improves
the reported error.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>