Change the Admitter interface's Admit func signature
Add a context parameter as a first parameter to the Admit function in
the Admitter interface.

Change all the implementations of the Admitter interface to take
a context as the first parameter.

Fix all the unit tests that are using the interface implementations, to
send a context as the first parameter of the Admit method.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>