validating webhooks: Unexport the Admitter interface
Currently, the `Admitter` interface is exported and
the following structs (validating webhooks [1]) are "explicitly declaring" that
they implement the `Admitter` interface:
- InstancetypeAdmitter
- ClusterInstancetypeAdmitter
- PreferenceAdmitter
- ClusterPreferenceAdmitter

This explicit declaration makes them dependent on the following package:
"kubevirt.io/kubevirt/pkg/util/webhooks/validating-webhooks"

Unexport the `Admitter` interface in order to decouple
usage from implementation.

In case one of the admitters will accidentally stop implementing
the `admitter` interface, the `Serve` [2] function will not accept it
as an admitter and the compilation will break.

This change is also aligned with the `mutator` [3] interface which
represents mutating webhooks [4].

[1] https://kubernetes.io/docs/reference/access-authn-authz/admission-controllers/#validatingadmissionwebhook

[2] https://github.com/kubevirt/kubevirt/blob/b5d3131f96a965e1cbd0e72d1e303146cdf2d5e1/pkg/util/webhooks/validating-webhooks/validating-webhook.go#L58

[3] https://github.com/kubevirt/kubevirt/blob/3bb20aa5d029cd667057d6445a07a0185d482b01/pkg/virt-api/webhooks/mutating-webhook/mutating-webhook.go#L40

[4] https://kubernetes.io/docs/reference/access-authn-authz/admission-controllers/#mutatingadmissionwebhook

Signed-off-by: Orel Misan <omisan@redhat.com>