Move multusNetworkAnnotationPool under pkg/network
As part of the effort to consolidate sig-network's
code under sig-network's ownership,
move `multusNetworkAnnotationPool` under the `multus` package.

Also rename it to `NetworkAnnotationPool` in order for it
to be exported, and due to the fact that it is now
used with the `multus` package prefix.

This is done in order to enable moving logic that uses
`multusNetworkAnnotationPool` under pkg/network
in the following commits.

Signed-off-by: Orel Misan <omisan@redhat.com>