net, netpod: Include the VMI UID in the nmstate dump logs
The netpod nmstate dumps about the pod network namespace current and
desired states are useful in troubleshooting the network setup of
virt-launcher pods if needed. It is also suppose to be a one-time
execution, making the logging at "info" level reasonable.

This change fixes a missing reference to a VMI identifier.
A custom logger with the VMI object is now used.

Signed-off-by: Edward Haas <edwardh@redhat.com>