tests/instancetype/upgrade: Replace Update with Patch
It is common practice on our e2e tests to use Update k8s API in order to
run the test. In order to ensure the change was executed, these also are
wrapped with Eventually/RetryOnConflict.
The issue is that most often tests do not really need to use Update (for
an atomic change) and Patch would be quicker and more precises.

Replacing Update in the following functions:
- updateInstancetypeMatcher
- updatePreferenceMatcher

Signed-off-by: Ram Lavi <ralavi@redhat.com>