api: persist volume migration info in the VM status
The volume migration in the VM status reports the current volumes in
use. Upon a successful volume migration, the vm controller will update
the list.
The idea behind this implementation consists that the spec contains the
desired state, while the status describes the current state. If the 2
have a mistamatch, it means that either the update is still in progress
or something failed.
Once the 2 state match, then the update operation was either successful
or have been cancelled.

Signed-off-by: Alice Frosi <afrosi@redhat.com>