virt-operator: Remove stores from application
Stores and Informers are build at the application level and
used to initialize the controller.
In particular, Stores are simply passed as argument
to the controller.
There is no need to build the Stores in the application
since it can be done during the controller initialization
starting from the informers.

Signed-off-by: fossedihelm <ffossemo@redhat.com>