virt-ctrl, template: Decouple net annotations gen on migration
Currently, when the templateService renders a migration
target pod, there is logic in place to determine
whether the source pod has an ordinal network scheme,
and update it to the new naming scheme if necessary.

Introduce the `targetAnnotationsGenerator` interface
in order to decouple this logic.

Currently, no other SIG seems to need this capability,
thus, it is specific to SIG-network and not a slice.

Implement the interface in the network annotation generator.

Signed-off-by: Orel Misan <omisan@redhat.com>