Register user-defined-network (UDN) binding on kubevirt CR (#3034)
* api: Introduce PrimaryUserDefinedNetworkBinding FG enabler

Introduce a new enabler: PrimaryUserDefinedNetworkBinding.
The default value is false.
This commit only adds the relevant API and documentation.
The logic of this enabler will be added in later commits.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* hco-operator: Add primary UDN Binding image env var

Adding a new image that is currently created by kubevirt/kubevirt [0].

[0] https://github.com/kubevirt/kubevirt/pull/10425

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* util/cluster: Set HCO main() to fail if PrimaryUDNImageEnvV env var is not set

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* operands/kubevirt: Set Kubevirt CR new primary UDN Binding plugin

The ability to register network bindings to Kubevirt VMs was proposed
[0] in order to allow Kubevirt VMs to connect to network bindings and
gain advanced user defined network capabilities.
The configuration is added to the kubevirt CR (kubevirt feature gate* +
configuring the binding parameters).

This commit deploys and registers a primary user-defined-network (UDN)
binding to the kubevirt CR, so it could be used by kubevirt users.

* enabling the `NetworkBindingPlugins` kubevirt
feature-gate on kubevirt CR is not needed as it is already set by
default by HCO [1].

[0] https://github.com/kubevirt/community/pull/280
[1]
https://github.com/kubevirt/hyperconverged-cluster-operator/pull/2603

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* docs/cluster-configuration: Add primaryUserDefinedNetworkBinding FG

Signed-off-by: Ram Lavi <ralavi@redhat.com>

---------

Signed-off-by: Ram Lavi <ralavi@redhat.com>