console plugin/proxy: Change AntiAffinity rule to 'Preferred' instead of 'Required' (#3051)
Due to the (default setting of) rollingUpdate of the 'kubevirt-console-plugin' and 'kubevirt-apiserver-proxy' Deployments, when there is a change, a 3rd pod is being scheduled as a replacement.
If there is a nodeSelector placement config alongside the anti-affinity rules, the replacement pod won't get scheduled if there are less than 3 nodes labeled with the node selector label.
We'll use 'PreferredDuringSchedulingIgnoredDuringExecution' instead of 'RequiredDuringSchedulingIgnoredDuringExecution' to soften the affinity rule so no pod will get stuck at Pending.

Signed-off-by: Oren Cohen <ocohen@redhat.com>