[tests] introduce a decorator for Periodic tests
Introduce a decorator to label tests for being
executed only on periodic lanes.
Tests labelled as `PERIODIC` will only be executed
on test lanes containing periodic in their name so that
they are going to be skipped on pre-submit lanes.
This is especially useful for long running an not critical tests
and it will help containing CI execution time on open PRs.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>