tests/operator, test_id:6255: Refactor to use patch
It is common practice on our e2e tests to use Update k8s API in order to
run the test. In order to ensure the change was executed, these also are
wrapped with RetryOnConflict.
On this test case, The RetryOnConflict is miss-used so that conflicts
will not be retried (the Get should be inside the RetryOnConflict).

Refactoring test to use Patch, as this would be more precise.

Signed-off-by: Ram Lavi <ralavi@redhat.com>