Fix the monitoring flaky functest (#3057)
* functest: avoid panic on exit, when running multiple times

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Fix the monitoring flaky functest

The test modifies the KubeVirt CR and expects that the
kubevirt_hco_out_of_band_modifications_total metric will increased by 1.
However, this test is sometimes fails, when the actual result is greater
by 1 from the expected metric value.

The reason for that is that another test - labels test, also modifies
the KubeVirt CR. In cases where the labels test runs first, sometimes
prometheus takes some time to update the metric, and then, in the
monitoring test, when reading the value before test, we're getting the
value before it was modified by the labels test, and when reading the
value after the KubeVirt modification, we get the metric values for both
modifications.

This PR fixes the issue by using the CDI CR instead of the KubeVirt CR
in the labels test, so there is only one modification for the KubeVirt
CR during the functional test suite, and the test should always pass.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

---------

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>