Replace nodelabeller/capabilities with libvirtxml
Libvirtxml is a package used for parsing libvirt domain capabilities
xml.

It is currently used in KubeVirt extensively (see for a example recent PR
to refactor live-migration-source parsing of virtwrap using it [1]).

virt-handler is currently coupled to nodelabeller/capabilities, a
package which defines a struct for unmarshaling a libvirt
domain capabilities XML into, a structure which libvirtxml already
includes in the form of libvirtxml.Caps.

This commit replaces all instances of nodelabeller/capabilities
with libvirtxml and removes the no-longer used nodelabeller/capabilities
package. This includes:

1. virt-handler/vm.go
2. virt-handler/options.go
3. virt-handler/options_test.go

[1] https://github.com/kubevirt/kubevirt/pull/10847

Signed-off-by: Adi Aloni <adialoni101@gmail.com>