cleanup(operator.go): Use client instead of virtctl
Use the KubeVirt client instead of using virtctl through a
NewRepeatableVirtctlCommand where applicable in
tests/operator/operator.go.

The tests state that virtctl is used so that the subresources
for starting/stopping a VM are used. The KubeVirt client uses these
subresources too and if you look at the code of virtctl you see that it
does nothing different than calling the client like the tests do now.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>