watch: consider imperative hotplug with volume migration
Fix hotplug when volume migration feature gate is enabled. The
imperative hotplug changes the VM/VMI volume set, but at the same time
doesn't set any update volume strategy. The default behavior when the
strategy isn't set is to consider the update as a replacement.

This changes fixes the misbehavior comparing the VM volumes with the VMI
ones to extimate correctly the kind of update. In the case of the
imperative hotplug, the volumes are changed on the VMI for epehmeral
hotplugged volumes.

Signed-off-by: Alice Frosi <afrosi@redhat.com>