Increase migration metrics periodicity
Before this commit, migration data was reported with the domainstats collector
every 30 seconds. The data includes the migrated and remaining memory bytes,
the memory transfer rate and the dirty rate.

Such a long periodicity for an ephemeral job reduces visibility about the it's progress.
Users are were not capable of following the data migration in "real time", and, for most
VMs, the 30 seconds periodicity, caused the data to only be reported once during the job,
and sometimes only after it ended.

To increase the periodicity, we moved the migration stats to a separate collector.
This collector has an event handler to track changes in the VMI status, to start
a job to gather migration information when a new migration begins. This runs
every 5 seconds (configurable).

Signed-off-by: João Vilaça <machadovilaca@gmail.com>