Use string param instead of node in WithNodeSelectorFor
Currently, `WithNodeSelectorFor` takes the pointer to
a node to create the nodeSelector field of a vmi.
In most of the cases we just have the node name. Indeed,
we can see that a dummy node obj is passed to respect the
function signature.
The node name is enough to achieve what we want.

Signed-off-by: fossedihelm <ffossemo@redhat.com>