tests: Allow more matches when watching Linux console
The LinuxExpecter function watches the console for the beginning of the
Linux boot (in order to ensure that keystrokes will not be sent to the
firmware.  This is inherently racy because it's easy to miss the initial
"Linux version" string if the expecter is not connected to the console
quickly enough.  We can make it less racy by allowing other strings to
match.  Switch to ExpectSwitchCase and add a regex for systemd messages
which occur more frequently and throughout the boot process.

This does not completely eliminate the race but should help immensely.

Signed-off-by: Adam Litke <alitke@redhat.com>