Merge revision e2e tests into one
Currently we have two e2e tests: the first one that checks
that a revision is created when a vm starts and the second one
that checks that the revision is updated when a vm is restarted.

The second one includes all the steps of the first one, so there is
no reason to have both.

Signed-off-by: fossedihelm <ffossemo@redhat.com>