fix: correctly check dry-run requests
In the pod eviction admitter some dry run requests were not caught
due to KubeVirt only checking the dry-run field in the admission review
but not in the object contained in the admission review.

This was causing dry run requests made through kubectl drain using the
'--dry-run=server' flag to be processed as non dry-run requests.

It's possible that this is Kubernetes bug but it should be anyway fixed
fixed in the meantime.

Signed-off-by: Antonio Cardace <acardace@redhat.com>