cleanup(tests,vm_test): Use client instead of virtctl
Use the KubeVirt client instead of using virtctl through a
NewRepeatableVirtctlCommand where applicable in tests/vm_test.go.

Internally virtctl uses the same client call, so using virtctl does not
provide any additional value.

There are more test cases in this file using virtctl which are not
changed in this commit because they need to be moved to
subresource_api_test.go or should be dropped entirely.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>