cleanup(tests,vm): Use client instead of virtctl
Use the KubeVirt client instead of using virtctl through a
NewRepeatableVirtlCommand where applicable in tests/vm_test.go.

This restructures the RunStrategy related VM tests to work without
virtctl and also consolidates the tests where possible to avoid
duplication of test code.

The newVirtualMachineWithRunStrategy helper is dropped and instead the
tests create VMs with the help of libvmi now. This change allows to use
a guestless VM in most of the tests.

Remove the need to run test '"Should leave a failed VMI using
RunStrategyOnce' in serial, by only killing the qemu process
of a single VMI.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>