Fix disk expansion logic
This commit fixes disk expansion logic in Kubevirt so, instead of taking into account the requested disk capacity, we actually calculate the usable size instead.

The usable size provided by the storage provisioner can differ from the one requested in the PVC capacity. This can lead to unwanted behavior when expanding the disk to the requested size.

Signed-off-by: Alvaro Romero <alromero@redhat.com>