migration test: remove duplicate dot import
kubevirt.io/kubevirt/tests/framework/matcher pkg was
imported twice: one with dot and one without it.

The dot import is a bad practice, so let's use
the named one.

Also renamed `libvirtDomainIsPersistent` and `libvirtDomainIsPaused`
respectively to `isLibvirtDomainPersistent` and `isLibvirtDomainPaused`
to better describe that the returned value is a boolean.

Signed-off-by: fossedihelm <ffossemo@redhat.com>