api, preference: Add PreferredEfi to API
The introduction of PreferredUseEfi and PreferredUseSecureBoot with the
original flavors API was a mistake. PreferredUseEfi in particular was
shortsighted and restrictive, being unable to express any newly
introduced attributes of the EFI struct within the core API.

This change introduces PreferredEfi as an eventual replacement for these
fields allowing us to express all EFI related attributes easily within a
preference.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>