node-labeller: Move TSC counter validation to Run()
Previously, TSC counter validation was being asserted during
prepareLabels and if the result was false, it would log an error as the
node-labeller failed to obtain the TSC counter value.

This is redundant as the TSC counter values don't change after
initialization therefore the assertion and log message were moved to
the node-labeller startup instead.

Signed-off-by: Adi Aloni <adialoni101@gmail.com>