tests/restore: add restore with vm source namespace doesnt exist
This is a scenario that would have failed before this PR
because the auth check on the restored dv would have happened
and failed and so the restore would fail to restore dvs and be stuck
on error.
Now with the new changes the check is skipped and the restore completes
successfully.
Also fixed a bug in the test where the VM was created in the source
namespace instead of a different namespace (the test checks for
cross namespace).

Signed-off-by: Shelly Kagan <skagan@redhat.com>