VMIController: Remove Expectations from VMI controller
Remove the Expectations usage from the VMI controller.
It was used to prevent multiple launcher pods from being
created when informers weren't synced, but this can be
handled more simply with a List API call.
The Expectations package, copied from Kubernetes in 2016,
is complex and outdated.
Removing it will simplify and clarify the KubeVirt codebase.

Signed-off-by: bmordeha <bmordeha@redhat.com>