usbredit: Don't assume order for client
As we spawn all clients in parallel we can't be
sure what goroutine will run sooner than others.
It well could be that last goroutine is running before
the first.

Note: There still seem to be issue with the test as the
code suggest that return value should be http status
code 503 and not 500.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>