VMRestore test: test to validate cleanup of vm after restore failure
The test uses existing test pipeline to get to a point
we see vm cant be start when a restore is failed using
a dummy webhook.
Changed the existing test to be a table which adds
an option to choose the way we cause the vm to start again,
either by deleting the webhook or deleting the vmrestore.
As the bug for this PR aims to fix, incase we are stuck
in a vmrestore which cant complete, if we delete it we want
to be able to start the vm after the vmrestore deletion.

Signed-off-by: Shelly Kagan <skagan@redhat.com>