fix(tests,virtctl): Partially revert cleanups
This partially reverts the cleanups of
be4d5584723cf9205976d37b6c9753285c02ee5c, because the now reverted
changes invalidate the results of the changed tests.

For the tests to have a meaningful result, the actual virtctl binary
needs to run (CreateCommandWithNS) instead of running the virtctl
command as part of the test binary (NewRepeatableVirtctlCommand).

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>