VMRestore: skip return reconcile in case of target status update
Since we changed from using statusupdater to
updateStatus we can get the updated vm instead of existing
the reconcile loop and waiting for it to get updated and then
reconcile again.
This is valid both in update target in progress
and update target restore is done.
Updated unit tests accordingly (since it saves a reconcile loop
no need for 2 unit tests for each scenario we do both in one loop)

Signed-off-by: Shelly Kagan <skagan@redhat.com>