vol-mig: verify required manual recovery flow
Verify restart and stop/start operation.
The test should refuse to restart the VM if the volume migration is in
progress. then adds the verification of the manual recovery after the
volume set restoration.

First, it verifies the VM start fails with an error message because of
the inconsistent state of the volume. Then, after the volume set fix,
the VM can be successfully restarted.

Signed-off-by: Alice Frosi <afrosi@redhat.com>