cleanup(tests,virtctl): Drop virtctl expose functests
The functional tests are testing functionality that is outside of the
scope of this command. The command is only creating a Service, while the
functional tests are asserting that basic network functionality of
Kubernetes is behaving correctly.

By ensuring the command creates Services as expected, all important test
cases in the scope of the command are already covered.

Therefore the the functional tests of virtctl expose are dropped.

The helper generateHelloWorldServer is moved closer to its remaining use
in bindingplugin_slirp.go. The no longer used helpers in libnet are
removed.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>