Bump OPM to v1.47.0 (#3095)
From OCP 4.17, the index image should contain `olm.csv.metadata` instead
of `olm.bundle.object`, as was in previous versions.

the few latest OPM didn't support this, but from OPM 1.47, we can use
the `--migrate-level=bundle-object-to-csv-metadata` flag to create the
desired FBC index image.

This PR bumps OPM to the latest version and add the
`--migrate-level=bundle-object-to-csv-metadata` flag, when running it.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>