VMRestore: fix possible miss update of dvname on vmrestore restores
This miss of update was spotted in recent tests flakes,
where we saw this variable missing in the test, while everything
else works fine.
Looking into how this can happen I noticed the update of this
variable in problematic places that with possible failure
of the vmrestore status then we might not ever get to update it later.
This commit fixes that by moving the update to before we generate
the restored vm, making sure we have a dedicated function that makes
sure this update is not missed.
Added UT to make sure this is happening and we have the flaky tests for
verification of this fix.

Signed-off-by: Shelly Kagan <skagan@redhat.com>