Remove wrong TODO about checking node allocated resources
The old TODO suggests to use the node allocated resources
in a wrong way, as if it reports the currently available
resources on a node.

This is a mistake since the allocated resources is intended
to report the total allocated resources of a node, which
is the node's capacity minus the reserved resources that
are being configured in kubelet.

The TODO points to a k8s issue which is closed due to
the fact that it's not a bug but intended behavior:
https://github.com/kubernetes/kubernetes/issues/64691

Signed-off-by: Itamar Holder <iholder@redhat.com>