vms-admitter.go: Remove TODO about assuming PVC.name equals DV.name
The assumption that the PVC and DV have the same name
is anyways spread across the repository and is a "fact
on the ground". If this assumption would change in
the future, it will be needed to be addressed in
many places and not just in here, so the TODO is
redandunt.

Signed-off-by: Itamar Holder <iholder@redhat.com>