virt-config: Graduate downwardMetrics to feature KV lifecycle
Enables the field: `spec.downwardMetrics: {}`. The configurable is
disabled by default and it brings the option to the user to enable it
cluster-wide if desired. Moreover, it adds checks to block a VM starting
if the feature is out of sync and runtime checks to inform the user when
the feature is out of sync too.

Signed-off-by: Javier Cano Cano <jcanocan@redhat.com>