[virt-tail] Replace watchFS polling with PollUntilContextTimeout
When virt-tail was first introduced, the logic for polling the
directory's existence during watchFS() setup was using a for loop with
sleep.

It was suggested then to use apimachinery's
PollUntilContextTimeout instead to simplify the logic[1] but the
implementation was postponed as the apimachinery version of KubeVirt at
the time didn't have PollUntilContextTimeout.

[1] https://github.com/kubevirt/kubevirt/pull/10110/files#r1341050446

Signed-off-by: Adi Aloni <adialoni101@gmail.com>